fixed int => long conversions on stream load

+++ src/NStore/Aggregates/IRepository.cs
@@ -17,13 +17,13 @@ namespace NStore.Aggregates
         
         Task<T> GetById<T>(
             string id,
-            int version,
+            long version,
             CancellationToken cancellationToken
         ) where T : IAggregate;       
         
         Task<T> GetById<T>(
             string id,
-            int version
+            long version
         ) where T : IAggregate;
 
         Task Save<T>(
